refactor: decomposition greeting and goodbye to dialog
