+++ src/index.js
@@ -1,23 +1,7 @@
 import os from 'os';
 import { App } from './app.js';
+import { createDialog } from './dialogue.js';
 
-const greeting = (username) => {
-  console.log(`Welcome to the File Manager, ${username}!`);
-};
-
-const goodbye = (username) => {
-  console.log(`Thank you for using File Manager, ${username}, goodbye!`);
-};
-
-let username = 'Anonymous';
-const args = process.argv.slice(2);
-const lastArg = args.at(-1);
-
-if (lastArg && lastArg.includes('--username')) {
-  username = lastArg.replace('--username=', '').trim() || 'Modest user';
-}
-
-greeting(username);
-process.on('exit', () => goodbye(username));
+createDialog();
 const app = new App(os.homedir());
 await app.start();
